\name{copyToClipboard}
\alias{copyToClipboard}
\title{Copies the variable to the clipboard
For data frames, writes as table (with separator tab), everything else directly.}
\usage{
  copyToClipboard(variable)
}
\description{
  Copies the variable to the clipboard For data frames,
  writes as table (with separator tab), everything else
  directly.
}

