% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{DataElement-class}
\alias{DataElement}
\alias{DataElement-class}
\title{Class representing a reference class DataElement that holds both basic
data information and settings information.}
\description{
This is a reference class so can be passed around the easily GUI for storing information.
}

