% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{dfv-package}
\alias{dfv}
\alias{dfv-package}
\title{dfv package}
\description{
Data Frame Viewer
}
\details{
Simple user interface to facilitate getting started with using R for data processing. Illustrates how to import data from Excel, melt data frames into plottable format, add additional information to the data and plot it using ggplot. Provides an easy system to keep track of multiple plots and save them in PDF format. Always shows the actual code that is executed to process or plot the data so users can experiment with changing the code directly and copy it to make their own data processing pipeline independent of this GUI.

The Data Frame Viewer can be run without any configuration for interacting with the data frames in any workspace.
}
\examples{
\dontrun{\\code{dfv.start()}}
}
\author{
Sebastian Kopf
}
\seealso{
\code{\link{dfv.start}}
}

