% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dfv.dev}
\alias{dfv.dev}
\title{Development launch, stores the module in object \code{dfv_test} and has debug messages turned on.}
\usage{
dfv.dev()
}
\description{
Development launch, stores the module in object \code{dfv_test} and has debug messages turned on.
}

