% Generated by roxygen2 (4.0.0): do not edit by hand
\name{generateCode.singleColumnMultiplot}
\alias{generateCode.singleColumnMultiplot}
\title{Generate the code for a multiplot of a single column}
\usage{
generateCode.singleColumnMultiplot(module, column)
}
\description{
Generate the code for a multiplot of a single column
}

