% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{GraphicsNotebook-class}
\alias{GraphicsNotebook}
\alias{GraphicsNotebook-class}
\title{Class implementing a fully flexible graphics notebook as a GuiElement}
\description{
Class implementing a fully flexible graphics notebook as a GuiElement
}

