% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{GraphicsNotebookTab-class}
\alias{GraphicsNotebookTab}
\alias{GraphicsNotebookTab-class}
\title{Class implementing a fully flexible graphics notebook as a GuiElement
FIXME: consider implementing such that it doesn't have to be tied to a GraphicsNotebook (stand alone wrapper for graphics devices)
FIXME: implement multi graphics support here or in separate class (e.g. MultiGraphNotebookTab)
       for multigraphics, use definition scheme similar to layout in multiplot}
\description{
Class implementing a fully flexible graphics notebook as a GuiElement
FIXME: consider implementing such that it doesn't have to be tied to a GraphicsNotebook (stand alone wrapper for graphics devices)
FIXME: implement multi graphics support here or in separate class (e.g. MultiGraphNotebookTab)
       for multigraphics, use definition scheme similar to layout in multiplot
}

