% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{GuiElement-class}
\alias{GuiElement}
\alias{GuiElement-class}
\title{A data frame that is linked to an element of the user interface (which can contain widgets and other GuiElements)}
\description{
A data frame that is linked to an element of the user interface (which can contain widgets and other GuiElements)
}

