% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{GuiModule-class}
\alias{GuiModule}
\alias{GuiModule-class}
\title{A module is a GuiElement that has a whole user interface associated with it.}
\description{
A module is a GuiElement that has a whole user interface associated with it.
}

