% Generated by roxygen2 (4.0.0): do not edit by hand
\name{showInfo}
\alias{showInfo}
\title{show an info message
type - styling of the mssage, info, error, question, warning are the standard ones
timer - time in seconds until message disappears automatically
okButton - whether there is an ok button or not}
\usage{
showInfo(gui, module, msg, type = "question", timer = 2, okButton = TRUE)
}
\description{
show an info message
type - styling of the mssage, info, error, question, warning are the standard ones
timer - time in seconds until message disappears automatically
okButton - whether there is an ok button or not
}

