\name{showInfo}
\alias{showInfo}
\title{show an info message
type - styling of the mssage, info, error, question, warning are the standard ones
timer - time in seconds until message disappears automatically
okButton - whether there is an ok button or not}
\description{
  show an info message type - styling of the mssage, info,
  error, question, warning are the standard ones timer -
  time in seconds until message disappears automatically
  okButton - whether there is an ok button or not
}

